[kotlin] Move KotlinAstUtil
