+++ pmd-kotlin/src/test/java/net/sourceforge/pmd/lang/kotlin/ast/internal/KotlinAstUtilTest.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.kotlin.util;
+package net.sourceforge.pmd.lang.kotlin.ast.internal;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
